Download Pre-Defined collection

Download a predefined collection of time series.

Usage

get_collection(collection_name, api_key = NULL, show_progress = FALSE)

Arguments

collection_name

The name of the collection you wish to download. For a list of available collections, go to ["list_available_collections"]

api_key

character Your API key. An API key is needed if accessing non-public timeseries. If downloading public timeseries parameter can be omitted.

show_progress

If collection to true, shows a progress bar of the data being downloaded.

Examples

get_collection("ds_kmi_mixed_freq",show_progress = TRUE)
#>   |                                                                              |                                                                      |   0%  |                                                                              |===================================                                   |  50%  |                                                                              |======================================================================| 100%
#> $ch.kof.ie.retro.ch_total.ind.d11
#>               Jan          Feb          Mar          Apr          May
#> 1992                                                                 
#> 1993 -22.14840787           NA           NA -20.29511378           NA
#> 1994  -7.41034205           NA           NA  -4.61507289           NA
#> 1995  -8.72226963           NA           NA -13.03854294           NA
#> 1996 -21.94780414           NA           NA -22.96136373           NA
#> 1997 -21.49035407           NA           NA -15.49167067           NA
#> 1998  -6.94196563           NA           NA  -7.29393935           NA
#> 1999  -3.25047769           NA           NA  -5.84398752           NA
#> 2000   1.85672404           NA           NA   9.27109443           NA
#> 2001  10.24698605           NA           NA   5.32583420           NA
#> 2002 -13.64518032           NA           NA -13.90175298           NA
#> 2003 -18.86139147           NA           NA -23.27780263           NA
#> 2004 -11.73267750           NA           NA  -8.77530401           NA
#> 2005  -7.19705448           NA           NA  -6.72642170           NA
#> 2006   0.02316601           NA           NA   3.91056738           NA
#> 2007   8.81009024           NA           NA   9.87061906           NA
#> 2008  10.13563134           NA           NA   9.16176753           NA
#> 2009 -10.34270726           NA           NA -16.56508118           NA
#> 2010  -3.13974005           NA           NA   1.19834036           NA
#> 2011   6.80560464           NA           NA   6.73507952           NA
#> 2012  -3.28838947           NA           NA  -1.39806391           NA
#> 2013  -1.39753734           NA           NA  -0.94324606           NA
#> 2014   1.01356574           NA           NA   1.31990403           NA
#> 2015  -3.20424240           NA           NA  -7.01680171           NA
#> 2016  -4.38733838           NA           NA  -3.99381013           NA
#> 2017  -2.18753943           NA           NA  -0.16317101           NA
#> 2018   3.72560248           NA           NA   4.52962665           NA
#> 2019   4.78341565           NA           NA   5.64917826           NA
#> 2020   3.60758762           NA           NA -20.68962311           NA
#> 2021  -5.96074248           NA           NA  -0.08935280           NA
#> 2022  11.54463292                                                    
#>               Jun          Jul          Aug          Sep          Oct
#> 1992              -19.30354105           NA           NA -26.32200156
#> 1993           NA -17.77650504           NA           NA -12.25242696
#> 1994           NA  -3.57137468           NA           NA  -4.16494864
#> 1995           NA -16.86863629           NA           NA -20.05433318
#> 1996           NA -23.39492843           NA           NA -24.68011807
#> 1997           NA -14.45373588           NA           NA  -9.43511574
#> 1998           NA  -5.57129971           NA           NA  -5.62007010
#> 1999           NA  -3.30458619           NA           NA   0.53041507
#> 2000           NA  11.77944157           NA           NA  13.80480756
#> 2001           NA  -2.64528920           NA           NA -10.50721333
#> 2002           NA -15.33725809           NA           NA -18.91871511
#> 2003           NA -22.37221236           NA           NA -17.45016864
#> 2004           NA  -6.87168425           NA           NA  -6.70173448
#> 2005           NA  -6.40807337           NA           NA  -1.94410113
#> 2006           NA   6.10987064           NA           NA   7.34960941
#> 2007           NA  10.66709015           NA           NA  10.42366028
#> 2008           NA   6.29714222           NA           NA  -0.86106833
#> 2009           NA -13.47277669           NA           NA  -7.98707347
#> 2010           NA   5.07021553           NA           NA   7.37106631
#> 2011           NA   2.37066650           NA           NA  -2.59938065
#> 2012           NA  -0.85752064           NA           NA  -1.79454404
#> 2013           NA  -1.07385192           NA           NA   1.27922766
#> 2014           NA   0.67068061           NA           NA  -0.69251832
#> 2015           NA  -6.37609328           NA           NA  -5.59696755
#> 2016           NA  -3.36132865           NA           NA  -3.07613909
#> 2017           NA   0.21348251           NA           NA   2.90828739
#> 2018           NA   5.33675165           NA           NA   4.82099493
#> 2019           NA   4.47621615           NA           NA   1.49431072
#> 2020           NA -12.42029825           NA           NA  -6.82456037
#> 2021           NA   7.61809914           NA           NA   9.53406671
#> 2022                                                                 
#>               Nov          Dec
#> 1992           NA           NA
#> 1993           NA           NA
#> 1994           NA           NA
#> 1995           NA           NA
#> 1996           NA           NA
#> 1997           NA           NA
#> 1998           NA           NA
#> 1999           NA           NA
#> 2000           NA           NA
#> 2001           NA           NA
#> 2002           NA           NA
#> 2003           NA           NA
#> 2004           NA           NA
#> 2005           NA           NA
#> 2006           NA           NA
#> 2007           NA           NA
#> 2008           NA           NA
#> 2009           NA           NA
#> 2010           NA           NA
#> 2011           NA           NA
#> 2012           NA           NA
#> 2013           NA           NA
#> 2014           NA           NA
#> 2015           NA           NA
#> 2016           NA           NA
#> 2017           NA           NA
#> 2018           NA           NA
#> 2019           NA           NA
#> 2020           NA           NA
#> 2021           NA           NA
#> 2022                          
#> 
#> $kofbarometer
#>            Jan       Feb       Mar       Apr       May       Jun       Jul
#> 1991  79.72002  71.81318  75.07457  81.14532  82.10748  86.19379  88.48660
#> 1992 101.05948 103.23209 107.31012 100.79347 102.11429 102.25721  91.29009
#> 1993  96.92816 102.53944 104.10948  98.43463 106.04517 107.64759 106.01730
#> 1994 106.02652 111.99426 108.85419 116.71908 115.47188 110.37501 107.52145
#> 1995  92.55774  96.23581  94.71102  78.93890  89.04406  84.80928  84.16842
#> 1996  90.87637  86.51605  88.61169  97.62766  88.19480  96.31743 101.69950
#> 1997 110.98017 110.27556 109.82194 114.43760 113.42183 110.90675 112.43675
#> 1998 110.71520 103.89915 103.08242 102.19273 104.29741  99.45116 103.97089
#> 1999  96.71402  98.03052 100.15368 100.32839 105.98236 109.91911 108.32491
#> 2000 108.43110 108.26395 111.70161 109.89711 108.04771 108.33340 104.09083
#> 2001  95.99892  91.06629  86.13479  89.70476  83.66573  84.47871  87.81404
#> 2002  88.34145  95.19410 100.41111  97.12962 102.70207 100.21015  93.66135
#> 2003  96.10151  91.92267  91.16587  88.98716  89.17348  89.63200  97.68864
#> 2004 112.15751 113.60185 113.67804 115.25919 116.83423 113.22926 111.44238
#> 2005  97.10564  98.36620  97.44742  98.15283  96.11908  98.09322  99.08757
#> 2006 107.96510 109.33074 107.33269 108.14664 107.48757 108.69499 104.69359
#> 2007 104.00963 102.71559 104.68121 100.89636  98.97773  98.58660  99.64342
#> 2008  97.44669  92.94459  89.84362  89.99611  89.24867  86.70467  87.81888
#> 2009  65.77258  60.08210  62.24697  70.79740  79.73676  88.71075  92.95813
#> 2010 120.33924 131.44301 131.87266 128.06141 128.78412 123.57125 117.24575
#> 2011  98.86280  96.82948  98.16249  95.83302  95.07184  91.41901  85.03864
#> 2012  85.27232  88.74268  91.74636  93.07069  92.23562  93.12483  94.68657
#> 2013 106.92170 107.84237 106.39001 103.44231 101.38481 100.45905 105.86677
#> 2014 105.89806 103.30341 101.72642 103.15048 101.63136 101.19522  98.21855
#> 2015  93.89095  84.16439  85.34179  82.82366  94.24204  91.61848  90.80227
#> 2016  94.06265 102.77027 103.68953 104.63844 102.07790 104.07387 104.86324
#> 2017 103.60095 106.74150 107.44505 109.10973 108.66758 109.19187 107.85476
#> 2018 106.35701 102.92392  99.95290  99.31586  98.32276  96.38842  96.50916
#> 2019  94.12919  91.89710  94.00193  95.71078  92.39582  91.80025  94.19132
#> 2020  99.46512 100.63308  88.89461  56.19808  49.64002  62.94850  83.41661
#> 2021  96.33497 103.17571 119.48584 135.48195 143.70926 132.82158 130.20693
#> 2022 107.23520 105.34084  99.72106                                        
#>            Aug       Sep       Oct       Nov       Dec
#> 1991  90.55181  98.96853 103.25049 103.27702 105.21727
#> 1992  94.49415  92.59879  87.78372  95.61071  90.09072
#> 1993 110.18915 111.22775 108.38259 108.88424 113.19467
#> 1994 110.21824 111.85006 105.79839 110.60699 106.14378
#> 1995  83.67213  81.35920  79.85781  80.22949  80.59492
#> 1996  99.48595 106.47739 104.50588 108.21608 110.01970
#> 1997 112.17489 110.25355 118.08638 113.68064 108.75840
#> 1998  98.62445  95.78553  94.27444  90.54880  93.55992
#> 1999 111.51269 116.19225 114.47731 117.04888 115.92861
#> 2000 103.81386  98.76768 100.94561  98.73249  95.42500
#> 2001  86.42248  85.96333  78.32248  75.38834  80.39122
#> 2002  93.69065  93.69281  95.75936  97.95547  97.49544
#> 2003  99.32759 104.72737 108.66960 111.32748 112.97104
#> 2004 107.22247 103.56140 102.10870  96.10223  97.65053
#> 2005 100.34727 104.13141 105.25228 106.95583 107.51539
#> 2006 105.49889 102.83837 103.32700 106.10890 104.65731
#> 2007  99.39879  99.61102  98.72527  98.21441  97.84928
#> 2008  82.33394  80.35839  81.06127  70.01807  63.81892
#> 2009 103.27440 111.86453 109.02193 121.08731 125.45306
#> 2010 114.06641 109.58338 106.37554 104.32319 103.34844
#> 2011  79.23764  75.67185  76.77952  79.10999  82.07778
#> 2012 102.52423 105.15231 103.96014 104.61959 103.39599
#> 2013 103.31154 105.81013 107.61084 108.57098 107.44500
#> 2014  97.78610  96.77918  96.78486  95.62390  95.78197
#> 2015  96.01643  93.05400  93.02504  95.08677  94.93465
#> 2016 101.50608 103.38228 103.74987 101.68905 103.64799
#> 2017 104.08484 106.72188 107.15507 107.93663 107.05078
#> 2018  95.91990  96.28810  96.95780  95.27019  93.03929
#> 2019  93.02314  90.37754  92.57205  91.83516  95.53334
#> 2020 109.92999 112.88605 107.85259 106.21103 105.85571
#> 2021 113.15968 110.56733 109.99356 107.32081 107.10811
#> 2022                                                  
#>