Get Time Series form KOF Datenservice API

Download time series data from the KOF web API. To explore the available data and find the keys to series you are interested in, run start_key_explorer.

get_time_series(ts_keys, api_key = NULL, show_progress = FALSE)

Arguments

ts_keys

A vector of timeseries keys

api_key

Your API key. This is only needed if accessing non-public time series.

show_progress

If set to true, shows a progress bar of the data being downloaded. year-month-day, otherwise only year and month.

Examples

get_time_series("kofbarometer")
#> $kofbarometer
#>            Jan       Feb       Mar       Apr       May       Jun       Jul
#> 1991  79.71922  71.81188  75.07348  81.14461  82.10684  86.19341  88.46763
#> 1992 101.09468 103.21663 107.28031 100.81879 102.11428 102.26545  91.28101
#> 1993  96.92119 102.52581 104.14619  98.38338 106.05948 107.66533 106.00718
#> 1994 106.02921 111.98645 108.85922 116.73987 115.41973 110.36156 107.51390
#> 1995  92.57062  96.21946  94.72041  78.92196  89.09486  84.86867  84.17998
#> 1996  90.88146  86.53399  88.59819  97.62629  88.16120  96.30701 101.70463
#> 1997 110.99263 110.30061 109.86544 114.46027 113.42636 110.86635 112.41682
#> 1998 110.71344 103.80484 103.05683 102.18148 104.28983  99.48749 104.04059
#> 1999  96.72449  98.13075 100.20898 100.31947 105.94830 109.85229 108.27055
#> 2000 108.44128 108.26614 111.69036 109.92060 108.07232 108.37045 104.09835
#> 2001  95.99903  91.06940  86.19037  89.71480  83.66708  84.46555  87.81903
#> 2002  88.33697  95.17222 100.38260  97.11407 102.71597 100.18709  93.66323
#> 2003  96.10291  91.92193  91.18893  88.99548  89.17537  89.63152  97.68142
#> 2004 112.16291 113.62186 113.62739 115.26009 116.83739 113.25558 111.45068
#> 2005  97.09928  98.33921  97.46436  98.15903  96.11400  98.08475  99.07555
#> 2006 107.96068 109.33167 107.33081 108.13989 107.48148 108.67576 104.69397
#> 2007 104.01196 102.72274 104.71543 100.89624  98.99237  98.56928  99.65559
#> 2008  97.44611  92.94602  89.80205  89.99227  89.24181  86.72062  87.81500
#> 2009  65.76371  60.05948  62.25763  70.80152  79.73766  88.72922  92.95942
#> 2010 120.34166 131.42729 131.89630 128.05930 128.79428 123.55153 117.24261
#> 2011  98.86055  96.84560  98.16234  95.84015  95.05160  91.41073  85.04002
#> 2012  85.27981  88.73993  91.70507  93.06368  92.22757  93.16179  94.68456
#> 2013 106.91858 107.85211 106.41963 103.44433 101.38813 100.44346 105.86715
#> 2014 105.89878 103.29060 101.74102 103.14813 101.62012 101.17051  98.21739
#> 2015  93.88685  84.16824  85.32907  82.82171  94.23102  91.61559  90.80844
#> 2016  94.06374 102.76839 103.67142 104.63859 102.08231 104.10601 104.85082
#> 2017 103.60163 106.73367 107.44984 109.11103 108.68439 109.18317 107.86207
#> 2018 106.35597 102.92442  99.97659  99.31357  98.31953  96.36934  96.50332
#> 2019  94.12878  91.90972  94.01327  95.71738  92.42258  91.80643  94.19623
#> 2020  99.47852 100.63168  88.87351  56.20760  49.59863  62.93452  83.40583
#> 2021  96.34052 103.20913 119.53476 135.49026 143.71562 132.85060 130.21761
#> 2022 107.20910 105.03530                                                  
#>            Aug       Sep       Oct       Nov       Dec
#> 1991  90.62401  98.95474 103.24384 103.27103 105.18320
#> 1992  94.39286  92.63255  87.78409  95.51185  90.04833
#> 1993 110.27572 111.22957 108.40122 108.91420 113.22479
#> 1994 110.22834 111.84883 105.79468 110.62549 106.16212
#> 1995  83.66814  81.33611  79.85796  80.19870  80.54158
#> 1996  99.44368 106.48277 104.50523 108.21160 110.07308
#> 1997 112.17351 110.27691 118.08767 113.68079 108.74875
#> 1998  98.64973  95.70604  94.18400  90.48580  93.54094
#> 1999 111.51383 116.24408 114.52667 117.09594 115.95934
#> 2000 103.78734  98.74470 100.93496  98.73504  95.39252
#> 2001  86.42271  86.00361  78.33193  75.37685  80.38800
#> 2002  93.69436  93.67245  95.75205  97.95999  97.51552
#> 2003  99.32796 104.72788 108.66804 111.31325 112.98022
#> 2004 107.21541 103.57171 102.10832  96.11708  97.64548
#> 2005 100.34205 104.11653 105.25604 106.95805 107.51579
#> 2006 105.51482 102.84403 103.33353 106.10380 104.66927
#> 2007  99.39491  99.60097  98.71403  98.20614  97.84257
#> 2008  82.32404  80.37878  81.06624  70.02182  63.80236
#> 2009 103.29382 111.84298 109.02922 121.10474 125.46376
#> 2010 114.04948 109.59162 106.37764 104.32070 103.33905
#> 2011  79.25448  75.66737  76.77120  79.11796  82.07154
#> 2012 102.51767 105.15530 103.96991 104.59937 103.39891
#> 2013 103.31355 105.80818 107.60822 108.57235 107.45374
#> 2014  97.77226  96.77720  96.77996  95.61727  95.77137
#> 2015  95.98483  93.05183  93.01868  95.08779  94.92693
#> 2016 101.51586 103.37504 103.76483 101.69078 103.64955
#> 2017 104.07465 106.72670 107.14748 107.94249 107.04801
#> 2018  95.92786  96.28443  96.95787  95.27142  93.05640
#> 2019  93.04066  90.38132  92.57632  91.83887  95.55165
#> 2020 110.00135 112.94101 107.86164 106.24242 105.85243
#> 2021 113.16473 110.54196 109.99723 107.27405 107.07909
#> 2022                                                  
#>