List Available Set Names and Their Description

Sets are pre-defined lists of time series. Sets are the convenient alternative to concatenating many series in an URL.

Usage

list_available_collections(api_key = NULL)

Arguments

api_key

Your API key. This is only needed if accessing non-public sets

Value

A data frame with the names, descriptions and public availability of datasets. These sets can be downloaded via get_dataset.

Examples

list_available_collections()
#>                                                                  set_name
#> baro_vintages_monthly                               baro_vintages_monthly
#> bs_indicator                                                 bs_indicator
#> ch.kof.trsm.fcst.summer                           ch.kof.trsm.fcst.summer
#> ch.kof.trsm.fcst.winter                           ch.kof.trsm.fcst.winter
#> ch.kof.trsm.fcst.year                               ch.kof.trsm.fcst.year
#> ds_consensus_qtr                                         ds_consensus_qtr
#> ds_disp_forecasters_qtr                           ds_disp_forecasters_qtr
#> ds_globidx_2018_full_fx                           ds_globidx_2018_full_fx
#> ds_globidx.v2017                                         ds_globidx.v2017
#> ds_globidx.v2018                                         ds_globidx.v2018
#> ds_globidx.v2019                                         ds_globidx.v2019
#> ds_kmi_mixed_freq                                       ds_kmi_mixed_freq
#> ds_kof_baublatt_ausblick_fitted_qtr   ds_kof_baublatt_ausblick_fitted_qtr
#> ds_kof_baublatt_ausblick_nowcast_qtr ds_kof_baublatt_ausblick_nowcast_qtr
#> ds_kof_baublatt_ausblick_qtr                 ds_kof_baublatt_ausblick_qtr
#> ds.sbfi.pan                                                   ds.sbfi.pan
#> globalbaro_vintages                                   globalbaro_vintages
#> kof_baublatt_ausblick                               kof_baublatt_ausblick
#> kof_consensus_forecast                             kof_consensus_forecast
#> kof_consensus_forecast_count                 kof_consensus_forecast_count
#> kof_consensus_forecast_dates                 kof_consensus_forecast_dates
#> kof_consensus_forecast_max                     kof_consensus_forecast_max
#> kof_consensus_forecast_mean                   kof_consensus_forecast_mean
#> kof_consensus_forecast_median               kof_consensus_forecast_median
#> kof_consensus_forecast_min                     kof_consensus_forecast_min
#> kof_consensus_forecast_prob                   kof_consensus_forecast_prob
#> kof_consensus_forecast_stdev                 kof_consensus_forecast_stdev
#> ogd_ch.kof.barometer                                 ogd_ch.kof.barometer
#> ogd_ch.kof.bts_total                                 ogd_ch.kof.bts_total
#> ogd_ch.kof.esi                                             ogd_ch.kof.esi
#> ogd_ch.kof.globalbaro                               ogd_ch.kof.globalbaro
#> ogd_ch.kof.ie                                               ogd_ch.kof.ie
#> stringency_plus_web                                   stringency_plus_web
#> stringency_web                                             stringency_web
#>                                                                  set_description
#> baro_vintages_monthly                                     KOF Barometer vintages
#> bs_indicator                         KOF Geschäftslage Indikator und Subbranchen
#> ch.kof.trsm.fcst.summer                       KOF Tourism Forecast Summer Season
#> ch.kof.trsm.fcst.winter                       KOF Tourism Forecast Winter Season
#> ch.kof.trsm.fcst.year                                  KOF Tourism Forecast Year
#> ds_consensus_qtr                                          KOF Consensus Forecast
#> ds_disp_forecasters_qtr                                                         
#> ds_globidx_2018_full_fx                     KOF Globalization index 2018 vintage
#> ds_globidx.v2017                                                                
#> ds_globidx.v2018                            KOF Globalization index 2018 vintage
#> ds_globidx.v2019                                                     KOF globidx
#> ds_kmi_mixed_freq                                      KOF Main Indicators (KMI)
#> ds_kof_baublatt_ausblick_fitted_qtr          KOF-Baublatt-Ausblick fitted values
#> ds_kof_baublatt_ausblick_nowcast_qtr              KOF-Baublatt-Ausblick Realtime
#> ds_kof_baublatt_ausblick_qtr                               KOF-Baublatt-Ausblick
#> ds.sbfi.pan                             KOF innovation survey (PAN) set for SBFI
#> globalbaro_vintages                                    Global Barometer vintages
#> kof_baublatt_ausblick                                      KOF-Baublatt-Ausblick
#> kof_consensus_forecast                                    KOF Consensus Forecast
#> kof_consensus_forecast_count                        KOF Consensus Forecast Count
#> kof_consensus_forecast_dates                        KOF Consensus Forecast Dates
#> kof_consensus_forecast_max                        KOF Consensus Forecast Maximum
#> kof_consensus_forecast_mean                          KOF Consensus Forecast Mean
#> kof_consensus_forecast_median                      KOF Consensus Forecast Median
#> kof_consensus_forecast_min                        KOF Consensus Forecast Minimum
#> kof_consensus_forecast_prob                  KOF Consensus Forecast Probabilites
#> kof_consensus_forecast_stdev           KOF Consensus Forecast Standard Deviation
#> ogd_ch.kof.barometer                                                        <NA>
#> ogd_ch.kof.bts_total                                                        <NA>
#> ogd_ch.kof.esi                                                              <NA>
#> ogd_ch.kof.globalbaro                                                       <NA>
#> ogd_ch.kof.ie                                                               <NA>
#> stringency_plus_web                                  for graphics on the webpage
#> stringency_web                                              KOF Stringency Index
#>                                      is_public
#> baro_vintages_monthly                     TRUE
#> bs_indicator                              TRUE
#> ch.kof.trsm.fcst.summer                   TRUE
#> ch.kof.trsm.fcst.winter                   TRUE
#> ch.kof.trsm.fcst.year                     TRUE
#> ds_consensus_qtr                          TRUE
#> ds_disp_forecasters_qtr                   TRUE
#> ds_globidx_2018_full_fx                   TRUE
#> ds_globidx.v2017                          TRUE
#> ds_globidx.v2018                          TRUE
#> ds_globidx.v2019                          TRUE
#> ds_kmi_mixed_freq                         TRUE
#> ds_kof_baublatt_ausblick_fitted_qtr       TRUE
#> ds_kof_baublatt_ausblick_nowcast_qtr      TRUE
#> ds_kof_baublatt_ausblick_qtr              TRUE
#> ds.sbfi.pan                               TRUE
#> globalbaro_vintages                       TRUE
#> kof_baublatt_ausblick                     TRUE
#> kof_consensus_forecast                    TRUE
#> kof_consensus_forecast_count              TRUE
#> kof_consensus_forecast_dates              TRUE
#> kof_consensus_forecast_max                TRUE
#> kof_consensus_forecast_mean               TRUE
#> kof_consensus_forecast_median             TRUE
#> kof_consensus_forecast_min                TRUE
#> kof_consensus_forecast_prob               TRUE
#> kof_consensus_forecast_stdev              TRUE
#> ogd_ch.kof.barometer                      TRUE
#> ogd_ch.kof.bts_total                      TRUE
#> ogd_ch.kof.esi                            TRUE
#> ogd_ch.kof.globalbaro                     TRUE
#> ogd_ch.kof.ie                             TRUE
#> stringency_plus_web                       TRUE
#> stringency_web                            TRUE