translate_legacy_keys

Gives ts_keys for the given legacy keys. Since there is no one to one mapping from old to new keys, it is possible that multiple new keys are returned for an old key. Note: Not all possible new keys are returned. The keys returned by this function should serve as a starting point to find the specific new keys that are needed. For the key explorer to find the exact keys, see start_key_explorer.

translate_legacy_keys(legacy_keys, chunksize = 100)

Arguments

legacy_keys

A vector of legacy keys for which to find new keys

chunksize

integer to set chunksize in order to avoid URLs that are too long.

Value

A list of vectors of new ts_keys